asyncio successful... somewhat
managed to change the program to run asynchronously with asycnio and aiohttp, pretty big improvement compare to the previous synchronous version which took up to 45 minutes, the current version took as long as the `timeout` defined in line 55 in urlChecker().

i think there's still improvements to be made, and questions to answer, particular.

1. is this the best flow? main() --> checkLinks() --> urlChecker()?
2. is there really a need to have a list, `tasks`, and `asyncio.gather(*tasks)` in order to get asyncio to work?