Added some ... kind of user flow
It's not completed, but we now have a bare userflow.
I've also added export to json options, even thought with asyncio and aiohttp we're able to speed up our checks it is a good idea to have an export function so that we can bypass the checking altogther. also added some verbosity.

next thingies to do -
- create import json functionality
- create matplotlib getRespCodeStats()
- think about whether exportJSON() and exportDetails() ought to be part of bookmarkChecker() or not.